Create main.py
This is a basic structure of the code that we are going to implement
We would not use the built-in method in numpy to do Bounded Leplace Algorithm, instead we would implement it by using only math and random libraries
This program expects a COVID dataset which you can find in our GitHub link. The dataset was originally found on Kaggle.